lfa-reader/requirements.txt
numpy
RPi.GPIO
pyserial
